+++ app/forms/support/case_hub_migration_form_schema.rb
@@ -2,12 +2,8 @@ module Support
   #
   # Validate "create a new case" form details for a Hub migration case
   #
-  class CaseHubMigrationFormSchema < Dry::Validation::Contract
-    import_predicates_as_macros
-
-    config.messages.backend = :i18n
+  class CaseHubMigrationFormSchema < Schema
     config.messages.top_namespace = :case_migration_form
-    config.messages.load_paths << Rails.root.join("config/locales/validation/support/en.yml")
 
     params do
       required(:school_urn).value(:string)
@@ -25,7 +21,7 @@ module Support
     end
 
     # TODO: custom macro for phone number validation
-    rule(:phone_number).validate(max_size?: 11, format?: /(^$|^0\d{10,}$)/)
+    rule(:phone_number).validate(max_size?: 13, format?: /^$|^(0|\+?44)[12378]\d{8,9}$/)
 
     # TODO: custom macro using chronic
     rule(:estimated_procurement_completion_date).validate(format?: /(^$|^\d{2}\/\d{2}\/\d{4}$)/)