+++ config/routes.rb
@@ -56,11 +56,9 @@ Rails.application.routes.draw do
   #
   # FaF ------------------------------------------------------------------
   #
-  resources :faf, only: %i[index], path: "procurement-support", controller: :faf do
+  resources :fafs, except: %i[delete], path: "procurement-support", controller: :fafs do
     get :user_query, on: :collection, path: "user-query"
   end
-    
-
 
   #
   # Supported ------------------------------------------------------------------
@@ -83,9 +81,12 @@ Rails.application.routes.draw do
       resources :interactions, only: %i[new create show]
       scope module: :cases do
         resource :categorisation, only: %i[edit update]
+        resource :savings, only: %i[edit update]
+        resource :procurement_details, only: %i[edit update]
         resources :documents, only: %i[show]
         resource :resolution, only: %i[new create]
         resource :assignment, only: %i[new create]
+        resources :contracts, only: %i[edit update]
         resource :email, only: %i[create] do
           scope module: :emails do
             resource :type, only: %i[new create]