+++ spec/features/self-serve/support_requests/create_support_request_spec.rb
@@ -63,7 +63,7 @@ RSpec.feature "Create a new support request" do
     xit "asks for a phone number" do
       expect(find("span.govuk-caption-l")).to have_text "About you"
       expect(find("label.govuk-label--l")).to have_text "What is your phone number?"
-      expect(find("span.govuk-hint")).to have_text "Your phone number will be used by DfE's supported buying team to contact you about your request for help. It will not be used for marketing or any other purposes. You do not need to provide a phone number."
+      expect(find(".govuk-hint")).to have_text "Your phone number will be used by DfE's supported buying team to contact you about your request for help. It will not be used for marketing or any other purposes. You do not need to provide a phone number."
     end
 
     it "starts off at step 4" do
@@ -91,8 +91,8 @@ RSpec.feature "Create a new support request" do
         fill_in "support_form[phone_number]", with: "0123"
         click_continue
         expect(find("h2.govuk-error-summary__title")).to have_text "There is a problem"
-        expect(page).to have_link "Your phone number must have a minimum of 10 numbers and no spaces, starting with a zero", href: "#support-form-phone-number-field-error"
-        expect(find("span.govuk-error-message")).to have_text "Your phone number must have a minimum of 10 numbers and no spaces, starting with a zero"
+        expect(page).to have_link "Your phone number must have no spaces and begin with a 0 or +44, with a minimum of 10 and maximum 12 digits", href: "#support-form-phone-number-field-error"
+        expect(find(".govuk-error-message")).to have_text "Your phone number must have no spaces and begin with a 0 or +44, with a minimum of 10 and maximum 12 digits"
       end
 
       # step 1
@@ -100,8 +100,8 @@ RSpec.feature "Create a new support request" do
         fill_in "support_form[phone_number]", with: "11234567890"
         click_continue
         expect(find("h2.govuk-error-summary__title")).to have_text "There is a problem"
-        expect(page).to have_link "Your phone number must have a minimum of 10 numbers and no spaces, starting with a zero", href: "#support-form-phone-number-field-error"
-        expect(find("span.govuk-error-message")).to have_text "Your phone number must have a minimum of 10 numbers and no spaces, starting with a zero"
+        expect(page).to have_link "Your phone number must have no spaces and begin with a 0 or +44, with a minimum of 10 and maximum 12 digits", href: "#support-form-phone-number-field-error"
+        expect(find(".govuk-error-message")).to have_text "Your phone number must have no spaces and begin with a 0 or +44, with a minimum of 10 and maximum 12 digits"
       end
 
       # step 1
@@ -109,17 +109,17 @@ RSpec.feature "Create a new support request" do
         fill_in "support_form[phone_number]", with: "0123456789x"
         click_continue
         expect(find("h2.govuk-error-summary__title")).to have_text "There is a problem"
-        expect(page).to have_link "Your phone number must have a minimum of 10 numbers and no spaces, starting with a zero", href: "#support-form-phone-number-field-error"
-        expect(find("span.govuk-error-message")).to have_text "Your phone number must have a minimum of 10 numbers and no spaces, starting with a zero"
+        expect(page).to have_link "Your phone number must have no spaces and begin with a 0 or +44, with a minimum of 10 and maximum 12 digits", href: "#support-form-phone-number-field-error"
+        expect(find(".govuk-error-message")).to have_text "Your phone number must have no spaces and begin with a 0 or +44, with a minimum of 10 and maximum 12 digits"
       end
 
       # step 1
       it "(max size)" do
-        fill_in "support_form[phone_number]", with: "012345678901"
+        fill_in "support_form[phone_number]", with: "+4412345678901343"
         click_continue
         expect(find("h2.govuk-error-summary__title")).to have_text "There is a problem"
-        expect(page).to have_link "Your phone number can not have more than 11 digits", href: "#support-form-phone-number-field-error"
-        expect(find("span.govuk-error-message")).to have_text "Your phone number can not have more than 11 digits"
+        expect(page).to have_link "Your phone number can not have more than 12 digits", href: "#support-form-phone-number-field-error"
+        expect(find(".govuk-error-message")).to have_text "Your phone number can not have more than 12 digits"
       end
     end
   end
@@ -149,7 +149,7 @@ RSpec.feature "Create a new support request" do
         click_continue
         expect(find("h2.govuk-error-summary__title")).to have_text "There is a problem"
         expect(page).to have_link "You must select a school", href: "#support-form-school-urn-field-error"
-        expect(find("span.govuk-error-message")).to have_text "You must select a school"
+        expect(find(".govuk-error-message")).to have_text "You must select a school"
       end
     end
   end
@@ -184,10 +184,9 @@ RSpec.feature "Create a new support request" do
     it "infers the category from the chosen spec" do
       choose "1 September 2021"
       click_continue
-
       expect(find("span.govuk-caption-l")).to have_text "About your procurement"
       expect(find("label.govuk-label--l")).to have_text "How can we help?"
-      expect(find("span.govuk-hint")).to have_text "Briefly describe your problem in a few sentences."
+      expect(find(".govuk-hint")).to have_text "Briefly describe your problem in a few sentences."
     end
   end
 
@@ -217,12 +216,12 @@ RSpec.feature "Create a new support request" do
 
       expect(find("h2.govuk-error-summary__title")).to have_text "There is a problem"
       expect(page).to have_link "The type of procurement is required if you do not select an existing specification", href: "#support-form-category-id-field-error"
-      expect(find("span.govuk-error-message")).to have_text "The type of procurement is required if you do not select an existing specification"
+      expect(find(".govuk-error-message")).to have_text "The type of procurement is required if you do not select an existing specification"
 
       choose "Broadband"
       click_continue
 
-      expect(find("span.govuk-hint")).to have_text "Briefly describe your problem in a few sentences."
+      expect(find(".govuk-hint")).to have_text "Briefly describe your problem in a few sentences."
     end
   end
 
@@ -243,7 +242,7 @@ RSpec.feature "Create a new support request" do
 
       expect(find("h2.govuk-error-summary__title")).to have_text "There is a problem"
       expect(page).to have_link "You must tell us how we can help", href: "#support-form-message-body-field-error"
-      expect(find("span.govuk-error-message")).to have_text "You must tell us how we can help"
+      expect(find(".govuk-error-message")).to have_text "You must tell us how we can help"
     end
   end